Move "EnhancedRecipe" and "Count" types into "enhancedEntities" folder.
