+++ src/lib/domain/enhancedEntities/Count.ts
