+++ src/lib/helpers/undefinedHelpers.ts
@@ -1,7 +1,7 @@
-export function isNullOrUndefined(obj: unknown) {
+export function isNullOrUndefined(obj: unknown): boolean {
   return obj === undefined || obj === null;
 }
 
-export function isPresent(obj: unknown) {
+export function isPresent(obj: unknown): boolean {
   return obj !== undefined && obj !== null;
 }
