+++ src/routes/recipes/+page.svelte
@@ -3,6 +3,7 @@
   import { getAllCoffeeBeans } from "$lib/database/manageCoffeeBeans";
   import { getAllRecipes } from "$lib/database/manageRecipes";
   import { routes } from "$lib/domain/constants/routes";
+  import { createEnhancedRecipe, type EnhancedRecipe } from "$lib/domain/enhancedEntities/EnhancedRecipe";
   import type { CoffeeBeans } from "$lib/domain/entities/CoffeeBeans";
   import { Recipe } from "$lib/domain/entities/Recipe";
   import type { RecipesSortOrderEnum } from "$lib/domain/sort/RecipesSortOrderEnum";
@@ -10,7 +11,6 @@
     sortRecipesByTimestampDesc as byTimestampDesc,
     sortRecipesByTimestampDesc
   } from "$lib/domain/sort/sortRecipes";
-  import { createEnhancedRecipe, type EnhancedRecipe } from "$lib/helperTypes/EnhancedRecipe";
   import RecipeCard from "$lib/UI/domainComponents/cards/RecipeCard.svelte";
   import MyFab from "$lib/UI/domainComponents/FABs/AddRecipeFab.svelte";
   import Loading from "$lib/UI/domainComponents/lists/Loading.svelte";
